+++ resources/sys_patch/sys_patch_auto.py
@@ -9,6 +9,7 @@ import logging
 import plistlib
 import subprocess
 import webbrowser
+import hashlib
 
 from pathlib import Path
 
@@ -331,7 +332,7 @@ Please check the Github page for more information about this release."""
             logging.info("- Unable to determine if boot disk is removable, skipping prompt")
 
 
-    def install_auto_patcher_launch_agent(self):
+    def install_auto_patcher_launch_agent(self, kdk_caching_needed: bool = False):
         """
         Install the Auto Patcher Launch Agent
 
@@ -350,12 +351,16 @@ Please check the Github page for more information about this release."""
             self.constants.auto_patch_launch_agent_path:        "/Library/LaunchAgents/com.dortania.opencore-legacy-patcher.auto-patch.plist",
             self.constants.update_launch_daemon_path:           "/Library/LaunchDaemons/com.dortania.opencore-legacy-patcher.macos-update.plist",
             **({ self.constants.rsr_monitor_launch_daemon_path: "/Library/LaunchDaemons/com.dortania.opencore-legacy-patcher.rsr-monitor.plist" } if self._create_rsr_monitor_daemon() else {}),
+            **({ self.constants.kdk_launch_daemon_path:         "/Library/LaunchDaemons/com.dortania.opencore-legacy-patcher.os-caching.plist" } if kdk_caching_needed is True else {} ),
         }
 
         for service in services:
             name = Path(service).name
             logging.info(f"- Installing {name}")
             if Path(services[service]).exists():
+                if hashlib.sha256(open(service, "rb").read()).hexdigest() == hashlib.sha256(open(services[service], "rb").read()).hexdigest():
+                    logging.info(f"  - {name} checksums match, skipping")
+                    continue
                 logging.info(f"  - Existing service found, removing")
                 utilities.process_status(utilities.elevated(["rm", services[service]], stdout=subprocess.PIPE, stderr=subprocess.STDOUT))
             # Create parent directories