+++ resources/wx_gui/gui_entry.py
@@ -6,6 +6,7 @@ import logging
 
 from resources import constants
 from resources.wx_gui import (
+    gui_cache_os_update,
     gui_main_menu,
     gui_build,
     gui_install_oc,
@@ -24,6 +25,7 @@ class SupportedEntryPoints:
     INSTALL_OC = gui_install_oc.InstallOCFrame
     SYS_PATCH  = gui_sys_patch_start.SysPatchStartFrame
     UPDATE_APP = gui_update.UpdateFrame
+    OS_CACHE   = gui_cache_os_update.OSUpdateFrame
 
 
 class EntryPoint: